# docs/swagger.yaml
basePath: /
definitions:
  mygram-byferdiansyah_comment_utils.ResponseMessage:
    properties:
      data:
        example: the error explained here
        type: string
      status:
        example: fail
        type: string
    type: object
  mygram-byferdiansyah_comment_utils.User:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  mygram-byferdiansyah_image_utils.ResponseMessage:
    properties:
      data:
        example: the error explained here
        type: string
      status:
        example: fail
        type: string
    type: object
  mygram-byferdiansyah_image_utils.User:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
...
  utils.AddComment:
    properties:
      image_id:
        example: image-123
        type: string
      message:
        example: A comment
        type: string
    type: object
  utils.AddImage:
    properties:
      caption:
        example: A caption
        type: string
      image_url:
        example: https://www.example.com/image.jpg
        type: string
      title:
        example: A Title
        type: string
    type: object
...
  utils.AddedComment:
    properties:
      created_at:
        example: the created at generated here
        type: string
      id:
        example: here is the generated comment id
        type: string
      image_id:
        example: here is the generated image id
        type: string
      message:
        example: A comment
        type: string
      user_id:
        example: here is the generated user id
        type: string
    type: object
  utils.AddedImage:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
  utils.EditComment:
    properties:
      message:
        example: A new comment
        type: string
    type: object
  utils.EditImage:
    properties:
      caption:
        example: A new caption
        type: string
      image_url:
        example: https://www.example.com/new-image.jpg
        type: string
      title:
        example: A new title
        type: string
    type: object
...
  utils.EditedComment:
    properties:
      caption:
        type: string
      id:
        type: string
      image_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  utils.EditedImage:
    properties:
      caption:
        type: string
      id:
        type: string
      image_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  utils.GetedComment:
    properties:
      created_at:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/utils.Image'
      image_id:
        type: string
      message:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/mygram-byferdiansyah_comment_utils.User'
      user_id:
        type: string
    type: object
  utils.GetedImage:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/mygram-byferdiansyah_image_utils.User'
      user_id:
        type: string
    type: object
  utils.Image:
    properties:
      caption:
        type: string
      id:
        type: string
      image_url:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
  utils.ResponseDataAddedComment:
    properties:
      data:
        $ref: '#/definitions/utils.AddedComment'
      status:
        example: success
        type: string
    type: object
  utils.ResponseDataAddedImage:
    properties:
      data:
        $ref: '#/definitions/utils.AddedImage'
      status:
        example: success
        type: string
    type: object
...
  utils.ResponseDataEditedComment:
    properties:
      data:
        $ref: '#/definitions/utils.EditedComment'
      status:
        example: success
        type: string
    type: object
  utils.ResponseDataEditedImage:
    properties:
      data:
        $ref: '#/definitions/utils.EditedImage'
      status:
        example: success
        type: string
    type: object
...
  utils.ResponseDataGetedComment:
    properties:
      data:
        items:
          $ref: '#/definitions/utils.GetedComment'
        type: array
      status:
        example: success
        type: string
    type: object
  utils.ResponseDataGetedImage:
    properties:
      data:
        items:
          $ref: '#/definitions/utils.GetedImage'
        type: array
      status:
        example: success
        type: string
    type: object
...
  utils.ResponseMessageDeletedComment:
    properties:
      message:
        example: your comment has been successfully deleted
        type: string
      status:
        example: success
        type: string
    type: object
  utils.ResponseMessageDeletedImage:
    properties:
      message:
        example: your image has been successfully deleted
        type: string
      status:
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: ferdicompany@gmail.com
    name: ferdi
  description: MyGram is a free image sharing app written in Go. People can share,
    view, and comment images by everyone. Anyone can create an account by registering
    an email address and creating a username.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: MyGram By Ferdiansya
  version: "1.0"
paths:
  /comments:
    get:
      consumes:
      - application/json
      description: Get all comments with authentication user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataGetedComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Get all comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: create and create a comment with authentication user
      parameters:
      - description: Add Comment
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.AddComment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ResponseDataAddedComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Add a comment
      tags:
      - comments
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a comment by id with authentication user
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseMessageDeletedComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Delete a comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Edit a comment by id with authentication user
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Edit Comment
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.EditComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataEditedComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_comment_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Edit a comment
      tags:
      - comments
  /images:
    get:
      consumes:
      - application/json
      description: Get all images with authentication user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataGetedImage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Get all images
      tags:
      - images
    post:
      consumes:
      - application/json
      description: Create and create a image with authentication user
      parameters:
      - description: Add Image
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.AddImage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ResponseDataAddedImage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Create a image
      tags:
      - images
  /images/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a image by id with authentication user
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseMessageDeletedImage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Delete a image
      tags:
      - images
    put:
      consumes:
      - application/json
      description: Edit a image by id with authentication user
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      - description: Image
        in: body
        name: json
        required: true
        schema:
          $ref: '#/definitions/utils.EditImage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseDataEditedImage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mygram-byferdiansyah_image_utils.ResponseMessage'
      security:
      - Bearer: []
      summary: Edit a image
      tags:
      - images
...
securityDefinitions:
  Bearer:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"